Directory restructured to avoid name conflict during plugin load
