+++ README.md
@@ -2,6 +2,6 @@
 
 ## Installation
 
-1) Copy the python/fledge/plugins/north/azure directory to /usr/local/fledge/python/fledge/python/plugins/north/
+1) Copy the python/fledge/plugins/north/azure_iot directory to /usr/local/fledge/python/fledge/python/plugins/north/
 
-2) pip3 install -Ir python/requirements-azure.txt
+2) pip3 install -Ir python/requirements-azure_iot.txt
