--- VERSION.north.azure
@@ -1,2 +1,0 @@
-fledge_north_azure_version=2.1.0
-fledge_version>=2.1
