+++ VERSION.north.azure_iot
@@ -1,0 +1,2 @@
+fledge_north_azure_iot_version=2.1.0
+fledge_version>=2.1
