+++ docs/images/azure_iot.jpg
