--- python/fledge/plugins/north/azure/__init__.py
@@ -1,4 +1,0 @@
-# Using the Python Azure Device SDK for IoT Hub:
-# https://github.com/Azure/azure-iot-sdk-python
-from azure.iot.device.aio import IoTHubDeviceClient
-from azure.iot.device import Message, MethodResponse
